\name{dat.cleasby2012}
\docType{data}
\alias{dat.cleasby2012.1}
\alias{dat.cleasby2012.2}
\alias{dat.cleasby2012}
\title{ADD_TITLE}
\description{ADD_DESCRIPTION}
\usage{dat.cleasby2012}
\format{
The data frame dat.cleasby2012.1 contains the following columns:
\tabular{lll}{
\bold{Spp} \tab \code{factor} \tab ADD_DESCRIPTION \cr
\bold{SppID} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{PopID} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{N} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Weight} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Zr} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{Year} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Type} \tab \code{factor} \tab ADD_DESCRIPTION \cr
\bold{Within} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Extra} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{Difference} \tab \code{integer} \tab ADD_DESCRIPTION \cr
}
The data frame dat.cleasby2012.2 contains the following columns:
\tabular{lll}{
\bold{Species} \tab \code{factor} \tab ADD_DESCRIPTION \cr
\bold{ZE} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{NE} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{WE} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{ZD} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{ND} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{WD} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{ZW} \tab \code{numeric} \tab ADD_DESCRIPTION \cr
\bold{NW} \tab \code{integer} \tab ADD_DESCRIPTION \cr
\bold{WW} \tab \code{integer} \tab ADD_DESCRIPTION \cr
}
}
\details{ADD_DETAILS}
\source{ADD_REFERENCE}
\author{ADD_CONTRIBUTOR_NAME, \email{ADD_EMAIL}}
\examples{
### copy data into 'dat' and examine data
dat <- dat.cleasby2012.1
dat
}
\keyword{datasets}
\concept{ADD_CONCEPT}
